+++ bootique-jetty/src/test/java/io/bootique/jetty/unit/JettyApp.java
@@ -2,9 +2,9 @@ package io.bootique.jetty.unit;
 
 import com.google.inject.Module;
 import io.bootique.BQCoreModule;
+import io.bootique.BQRuntime;
 import io.bootique.jetty.JettyModule;
 import io.bootique.jetty.command.ServerCommand;
-import io.bootique.test.BQDaemonTestRuntime;
 import io.bootique.test.junit.BQDaemonTestFactory;
 import org.eclipse.jetty.server.Server;
 
@@ -21,8 +21,8 @@ public class JettyApp extends BQDaemonTestFactory {
         }, args);
     }
 
-    public BQDaemonTestRuntime start(Module config, String... args) {
-        Function<BQDaemonTestRuntime, Boolean> startupCheck = r -> r.getRuntime().getInstance(Server.class).isStarted();
+    public BQRuntime start(Module config, String... args) {
+        Function<BQRuntime, Boolean> startupCheck = r -> r.getInstance(Server.class).isStarted();
 
         return app(args)
                 .modules(JettyModule.class)